Some bugs fixed, Tests passed

+++ src/main/java/com/mphj/todo/models/request/PostTodoRequest.java
@@ -2,17 +2,12 @@ package com.mphj.todo.models.request;
 
 import com.mphj.todo.entities.Todo;
 import com.mphj.todo.entities.UserList;
-import org.hibernate.validator.constraints.Range;
-import org.springframework.beans.factory.annotation.Required;
-import org.springframework.lang.Nullable;
 
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Null;
 import java.util.List;
 
 public class PostTodoRequest {
 
-    public int lastUpdateTime;
+    public long lastUpdateTime;
 
     public List<Todo> todoList;
